Add "report issue" button in error component #911
