+++ browser/data-browser/src/components/ErrorLook.tsx
@@ -4,6 +4,7 @@ import { styled, css } from 'styled-components';
 import { FaExclamationTriangle } from 'react-icons/fa';
 import { Column } from './Row';
 import { CodeBlock } from './CodeBlock';
+import { Button } from './Button';
 
 export const errorLookStyle = css`
   color: ${props => props.theme.colors.alert};
@@ -18,6 +19,43 @@ export const ErrorLook = styled.span`
 export interface ErrorBlockProps {
   error: Error;
   showTrace?: boolean;
+}
+
+const githubIssueTemplate = (
+  message,
+  stack,
+) => `**Describe what you did to produce the bug**
+
+## Error message
+\`\`\`
+${message}
+\`\`\`
+
+## Stack trace
+\`\`\`
+${stack}
+\`\`\`
+`;
+
+/** Returns github URL for new bugs */
+export function createGithubIssueLink(error: Error): string {
+  const url = new URL(
+    'https://github.com/atomicdata-dev/atomic-server/issues/new',
+  );
+  url.searchParams.set('body', githubIssueTemplate(error.message, error.stack));
+  url.searchParams.set('labels', 'bug');
+
+  console.log('opening', url);
+
+  return url.href;
+}
+
+export function GitHubIssueButton({ error }) {
+  return (
+    <Button onClick={() => window.open(createGithubIssueLink(error), '_blank')}>
+      Report on Github
+    </Button>
+  );
 }
 
 export function ErrorBlock({ error, showTrace }: ErrorBlockProps): JSX.Element {
@@ -46,14 +84,6 @@ const ErrorLookBig = styled.div`
   border-radius: ${p => p.theme.radius};
   border: 1px solid ${p => lighten(0.2, p.theme.colors.alert)};
   background-color: ${p => p.theme.colors.bg};
-`;
-
-const Pre = styled.pre`
-  white-space: pre-wrap;
-  border-radius: ${p => p.theme.radius};
-  padding: ${p => p.theme.margin}rem;
-  background-color: ${p => p.theme.colors.bg};
-  font-size: 0.9rem;
 `;
 
 const BiggerText = styled.p`
